src/ConsumptionTaxLawFill_in_the_BlankQuestions/scraper.py
import requests
from bs4 import BeautifulSoup, Tag

class Scraper:
    def __init__(self, url):
        self.url = url

    def scrape(self):
        response = requests.get(self.url)
        soup = BeautifulSoup(response.content, 'html.parser')

        # ここでページから必要なデータを抽出します。
        # 例えば、特定のタグやクラスを持つ要素を探す等。
        # 以下は仮のコードです。
        questions = soup.find_all('p')
        answers = soup.find_all('div')

        # 問題と答えを辞書形式で返します。
        questions_answers = {}
        for question, answer in zip(questions, answers):
            if isinstance(questions, Tag) and isinstance(answer, Tag):
                questions_answers[question.text] = answer.text
        return questions_answers

    # https://codezine.jp/article/detail/12230
    #スクレイピング参考ブログ
